// src/app/core/models/webinar.model.ts
export interface Webinar {
    title: string;
    description: string;
    video_url: string;
    tags: string[];
    docs: [];
    totalDocs: number;
    date: any;
    is_paid: boolean;
    cost: number;
    status: any;
    image: string[];
    id: string;
    createdAt: string;
    updatedAt: string;
    slug: string; 
    limit: number; 
}